+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changes since v2.6
 
 ### Enhancements
 - Application search tips hidden behind question mark icon (#1767)
+- Redirect to login page when accessing an attachment link when logged out (#1590)
 
 ### Fixes
 - More robust email resending (#1750)